fix bugs with plugin parsing
If the plugin fails to parse it formerly would just exit silently and
all plugins would fail to install.

Also, if the plugin output anything on stderr it would try to parse it
and also fail.

This redirects stderr out to the terminal and shows the parsing error if
any.