remove hpg shorthand (#37)
it has caused us 2 major issues now, and it actually still didn't match
the existing ruby functionality. Also, it only worked for addons:create
but not addons:upgrade so we decided to kill this.

A better solution here would be a more generic aliasing functionality or
a way for add-ons to specify aliases we think.