Fix env cascading when invoking psql
Currently we override most PG* environment variables, but there's no
good reason for this--if the user has set something, we should not
require them to override this via the URL.

With this change, they still *can* if they need to do so explicitly,
but we shouldn't clobber valid various from the environment with empty
ones.

We keep PGSSLMODE as a setting users cannot override (to avoid
unencrypted connections to remote databases), but account for unset
hosts (which defaults to a local connection).