Introduce new `releases:output` command (#164)
* introduce new `releases:output` command

This command looks for the availability of an `output_stream_url` field,
which will be a busl stream, and fetchs the log data from there.

* use the got package provided with heroku-cli-util instead of requests

* simplify check on undefined/null

* rename getLatestRelease in rollback to findRollbackRelease

We use the same method name in release output, yet they do different
things. So we need to make that clearer.

* extract finding a single release into a lib file

We do very similar things in releases index, rollback and output to find
a single release.
Extract that into a generic method.

* Use pipe to process.stdout instead of cli.log & refactor into finder (#168)

* Use std-mocks and pipe to stream output

* Factor out shared behavior in info / output

* Replace request with get