feat(apps): support creating an app with initial features (#1040)
Internal users are able to create an app with an initial set of feature
flags. Currently this is only possible by crafting a custom `curl` or
`heroku api` command. We maintain a few scripts which duplicate much of
what `heroku apps:create` provides in order to create apps with flags in
a space, or in a different region, etc.

This adds a new hidden `features` flag which accepts a comma-delimited
list of feature flags to set when creating an app. If the user is not
authorized to create an app with features enabled, they are silently
ignored.

`features` matches the Platform API naming and terraform's
`heroku_app_feature` resource.