fix(pg-v5): fix pg:pull & pg:push database identifiers
This commit fixes handling of database identifiers that have forward
slashes (`/`) in them. Database identifiers such as `myapp/my_test_db`
is a valid database name (i.e. `createdb myapp/my_test_db` succeeds) in
Postgres.

Currently, due to the use of `split()` to fetch the database identifier
from the connection string, only `myapp` would be returned in the
example above.

To solve this, we use the `pathname` property of the parsed property and
strip the leading slash if present (as relative URIs will not have it).