fix: Fix redis:promote for private plans
If the user is attempting to promote a Redis add-on and there already
*is* a REDIS attachment *and* it's the last attachment for that
add-on, we need to first create an alternate attachment (since
creating a new attachment implicitly detaches what was there before,
and the API won't allow us to remove the last attachment implicitly
like this).

Our current approach for doing this is naïve and breaks with private
Redis plans (or really any Redis configuration that has additional
config vars) because we rely on counting config vars.

The right solution is probably to look at attachment count for the
existing add-on attached at REDIS explicitly, but as a quick fix we
just look for config vars ending in '_URL' (rather than the total
count of config vars).