Remove warnUsingOrgFlagInTeams usage
To show the warning it assumes that the `team` flag isn't set,
but v5 plugins automatically receive a `team` flag when a v5
command is configured with `wantsOrg`. Because this `team` flag
exists, the warning won't be shown.

The `team` flag is added via `oclif/plugin-legacy` for v5 plugins
to be used within oclif, specifically here:
 https://github.com/oclif/plugin-legacy/blob/6d62fbe435b202776692c0929413acc870021f79/src/index.ts#L167-L171