Harcode screen width for testing purposes for consistent test results
Tests can be an issue when using cli-ux features that use oclif/screen.
Depending on the width of the terminal the output might wrap or truncate.
To get around this we set the global `columns` value, which is checked by
oclif/screen. This specifically impacted tables and error message
wrapping. 140 characters appears to be a width that works for the tests
currently.

See https://github.com/oclif/screen