chore: don't install bats twice
`bats` is already installed via npm (this was introduced in [#1252][1]),
so we don't need to install it in the CircleCI build file. The CircleCI
build file was also installing the original bats, while the [`local`][2] and
[`run`][3] packages install a [fork from npm][4].

In addition, I switched the command inside the bats runners from `npx
bats` (which will download/install the latest binary if `yarn install`
was not already run), to `yarn bats`, which will always use the version
specified in the lockfile. This is more consistent with how we run other
commands anyway.

[1]: https://github.com/heroku/cli/pull/1252
[2]: https://github.com/heroku/cli/blob/62085d32f8fa3574bbe03bba707f1f576a9858bf/packages/local/bin/bats-test-runner.js
[3]: https://github.com/heroku/cli/blob/62085d32f8fa3574bbe03bba707f1f576a9858bf/packages/run/bin/bats-test-runner.js
[4]: https://www.npmjs.com/package/bats