build: add release scripts
Takes the steps outlined in https://salesforce.quip.com/aPLDA1ZwjNlW
and creates some scripts that automate some of the steps so we don't
have to figure this out manually every time.

Added two scripts:

* `scripts/create-release-pr`
  * Creates a new branch (so a pull request can be opened)
  * Bumps the version and generates new changelogs using `yarn lerna version`
  * Opens a pull request
* `scripts/publish-release`
  * Checks that the script is only run on the main branch (`master` at
      the time of this writing)
  * Pulls down the latest Git branches/tags from GitHub
  * Creates a new tag based off of the version in `lerna.json`
  * Pushes tag to GitHub
  * Publishes packages to NPM
  * Creates a GitHub release with the correct release notes of things
  that changed based off of Git commit history

The scripts require installation of the GitHub CLI, and will
error/message appropriately if the [GitHub CLI][1] is not installed or
the person running the script is not logged into GitHub.

[1]: https://cli.github.com/