docs(git): Fix a mistake in git:remote examples
The example of adding a new remote with a custom alias stated that the
remote URL was `https://git.heroku.com/example-staging.git`, but the
command was pointing to `https://git.heroku.com/example.git`.

I've updated the example to use the same remote URL as the previous
example, so that the only difference is the addition of the `--remote`
argument.