build: use --force with git fetch origin --tags
If someone has released the CLI locally, they'll have a tag name for the
release that gets pulled from origin. Using `--force` is the only way to
override the local tag with the tag from GitHub, which should be the
canonical source of truth anyway and not cause conflicts since it should
be the same tag.