fix(apps-v5): get pipeline stage from `pipeline_coupling`
[775c679dd069342074cfbb96fea887e72db1c4f7][1] removed the `pipeline`
object from the `info` object used to render the display in `apps:info`
to fix a deprecation warning. Unfortunately, it seems like this
introduced a regression as code that was referencing `info.pipeline` was
not updated to use `info.pipeline_coupling`, resulting in runtime
exceptions due to trying to access a property off of `undefined`.

Work Item: [W-8780887][2]

[1]: https://github.com/heroku/cli/commit/775c679dd069342074cfbb96fea887e72db1c4f7
[2]: https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000094UCiIAM/view