feat(pg-v5): add burst checks, json output to diagnose
Heroku Postgres has some plans that have burstable performance
characteristics
(https://devcenter.heroku.com/articles/heroku-postgres-production-tier-technical-characterization#performance-characteristics).

Whether or not a given resource is bursting, or at risk at throttling,
has previously not been available to end users. This commit changes
this, by providing burst data to the `pg:diagnose` backend for relevant
resources, and returning it in the report.

This commit also adds JSON output support, for when the table output is
not appropriate and to provide some interop with tools like `jq`.

Ref: GUS-W-8906686