fix(pg-v5): handle databases without config vars
In rare cases, we can have a database that is attached, but has no
config vars, as the database provisioning has not yet updated API with
the config vars needed. This leaves the attachment in a state where
`config_vars` is empty, and thus breaks `getConfigVarName`, even when
asking for an entirely separate database.

This commit fixes this and adds a regression test to make sure this
doesn't happen.

Ref: https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000097iB0IAI/view