fix(apps-v5): Improve UX of stack:set command output
Previously the `stack:set` command's output mentioned needing
to create a "new release" in order to complete the stack change.

However this is incorrect - a new build needs to be performed,
so the wording has been updated accordingly

In addition, the usage text has been updated so that it matches
the actual output from the command.

Closes GUS-W-8797527.