fix(pg-v5): Fix databasename for URLs with query element and be more consistent
For operations such as pg:psql, we have been assuming that the URL parsed path
method is enough to find the database name. This is incorrect when we pass
parameters such as sslmode. In general, this problem has been solved with
parsePostgresConnectionString but not everything uses it.

This fix removes any custom parsing we are doing of URLs and moves to using
this function everywhere to avoid inconsistency.

* util.databaseNameFromUrl
* util.getConnectionDetails
* the pg:copy resolver

Without this, pg:psql was failing on verify-full URL's but with this patch, it works fine.