Update heroku-20 to heroku-22
The Heroku CLI's help text refers to Heroku-20 in a few of the examples. eg:


"""

$ heroku stack:set heroku-20 -a myapp

Setting stack to heroku-20... done

You will need to redeploy myapp for the change to take effect.

Run git push heroku main to trigger a new build on myapp.

"""


This updates it to use Heroku-22 instead.


See:

https://github.com/heroku/cli/search?q=heroku-20

GUS-W-10349673