Use older ubuntu runner to use older deb compression defaults
ubuntu-latest points to Ubuntu 22.04 as of this commit, which may use zstd by default for compressing all or a part of deb packages in `dpkg --build` or `dpkg-deb --build`. Debian 11 cannot use zstd, as filed in #2240. So, revert to building on an older Debian/Ubuntu base to maintain compatibility with Debian 11 since it is still supported.