feat(spaces): migrate drains:{set,get} to spaces topic
This commit partially migrates the `drains:set` and `drains:get`
commands for `spaces` under the `spaces` topic, so they can be accessed
from `spaces:drains:set` and `spaces:drains:get`.

The rationale here is that these are `spaces`-specific commands, and
there is significant semantic overlap between `drains:add` and
`drains:set`. This makes it clearer that these commands are
`spaces`-specific.

In order to not break comaptibility with existing workflows, the old
`drains:{set,get}` commands will continue to work, but remain hidden.