Fix CPU ISA detection logic in standalone install script
On Linux, `uname -m` will report `aarch64` on 64-bit ARM CPUs. The previous logic in this script matched that against the expression `aarch*`, causing the script to wrongly select the 32-bit ARM architecture identifier, which gives a runtime error when launching the CLI due to the lack of 32-bit dynamic linker on the system (specifically, the error indicates that `ld-linux-armhf.so.3` is missing).

Signed-off-by: Zach Mullen <zach.mullen@kitware.com>